requirements.txt
fastapi
uvicorn
PyPDF2
google-generativeai
python-dotenv
python-multipart
